Split code in two files - BERT_tokenizer.hpp and BERT_tokenizer_impl.hpp
BERT_tokenizer.hpp declares the classes and functions to be used and BERT_tokenizer_impl.hpp defines these
