+++ src/mlpack/core/data/BERT_tokenizer.hpp
@@ -54,119 +54,17 @@ private:
     bool mDoLowerCase;
 };
 
+static bool isStripChar(const wchar_t& ch);
 
-std::wstring BasicTokenizer::cleanText(const std::wstring& text) const {
-    std::wstring output;
-    for (const wchar_t& cp : text)  {
-        if (isWhitespace(cp)) output += L" ";
-        else output += cp;
-    }
-    return output;
-}
+static std::wstring strip(const std::wstring& text);
 
+static std::vector<std::wstring> split(const std::wstring& text);
 
-bool BasicTokenizer::isWhitespace(const wchar_t& ch) const {
-    if (ch== L' ' || ch== L'\t' || ch== L'\n' || ch== L'\r') return true;
-    return false;
-}
+static std::vector<std::wstring> whitespaceTokenize(const std::wstring& text);
 
-bool BasicTokenizer::isPunctuation(const wchar_t& ch) const {
-    if (ch == '!' || ch == ',' || ch == ';' || ch == '.' || ch == '?' ||   
-       ch == '-' || ch == '\'' || ch == '\"' || ch == ':' || ch == '(' ||
-       ch == ')' || ch == '[' || ch == ']' || ch == '{' || ch == '}' ) return true;
-    
-    return false;
-}
+void convertToUnicode(const std::string &s, std::wstring &ws);
 
-BasicTokenizer::BasicTokenizer(bool doLowerCase=true) 
-    : mDoLowerCase(doLowerCase) {
-}
-
-std::vector<std::wstring> BasicTokenizer::runSplitOnPunc(const std::wstring& text) const {
-    size_t i = 0;
-    bool startNewWord = true;
-    std::vector<std::wstring> output;
-    while (i < text.size()) {
-        wchar_t ch = text[i];
-        if (isPunctuation(ch)) {
-            output.push_back(std::wstring(&ch, 1));
-            startNewWord = true;
-        }
-        else {
-            if (startNewWord) output.push_back(std::wstring());
-            startNewWord = false;
-            output[output.size() - 1] += ch;
-        }
-        i++;
-    }
-    return output;
-}
-
-static bool isStripChar(const wchar_t& ch) {
-    return stripChar.find(ch) != std::wstring::npos;
-}
-
-static std::wstring strip(const std::wstring& text) {
-    std::wstring ret =  text;
-    if (ret.empty()) return ret;
-    size_t pos = 0;
-    while (pos < ret.size() && isStripChar(ret[pos])) pos++;
-    if (pos != 0) ret = ret.substr(pos, ret.size() - pos);
-    pos = ret.size() - 1;
-    while (pos != (size_t)-1 && isStripChar(ret[pos])) pos--;
-    return ret.substr(0, pos + 1);
-}
-
-static std::vector<std::wstring> split(const std::wstring& text) {
-    std::vector<std::wstring>  result;
-    boost::split(result, text, boost::is_any_of(stripChar));
-    return result;
-}
-
-static std::vector<std::wstring> whitespaceTokenize(const std::wstring& text) {
-    std::wstring rtext = strip(text);
-    if (rtext.empty()) return std::vector<std::wstring>();
-    return split(text);
-}
-
-void convertToUnicode(const std::string &s, std::wstring &ws) {
-
-    std::wstring wsTmp(s.begin(), s.end());
-
-    ws = wsTmp;
-}
-
-void convertFromUnicode(const std::wstring &ws, std::string &s) {
-
-    std::string sTmp(ws.begin(), ws.end());
-
-    s = sTmp;
-}
-
-std::vector<std::wstring> BasicTokenizer::tokenize(const std::string& text) const {
-    std::wstring nText;
-    convertToUnicode(text, nText);
-    nText = cleanText(nText);
-
-    const std::vector<std::wstring>& origTokens = whitespaceTokenize(nText);
-    std::vector<std::wstring> splitTokens;
-    for (std::wstring token : origTokens) {
-        if (mDoLowerCase) {
-            std::string t_s; //Temporary variable used for converting character to lower.
-            std::string token_s;
-            //Convert to std::string format.
-            convertFromUnicode(token, token_s);
-            mlpack::util::ToLower(token_s, t_s);
-            std::wstring t;
-            //Convert to std::wstring format.
-            convertToUnicode(t_s, t);
-            token = t;
-        }
-        const auto& tokens = runSplitOnPunc(token);
-        splitTokens.insert(splitTokens.end(), tokens.begin(), tokens.end());
-    }
-    return whitespaceTokenize(boost::join(splitTokens, L" "));
-}
+void convertFromUnicode(const std::wstring &ws, std::string &s);
 
 class WordpieceTokenizer {
 public:
@@ -178,49 +76,6 @@ private:
     std::wstring mUnkToken;
     size_t mMaxInputCharsPerWord;
 };
-
-WordpieceTokenizer::WordpieceTokenizer(const std::shared_ptr<Vocab> vocab, const std::wstring& unkToken, size_t maxInputCharsPerWord)
-    : mVocab(vocab),
-    mUnkToken(unkToken),
-    mMaxInputCharsPerWord(maxInputCharsPerWord) {
-}
-
-std::vector<std::wstring> WordpieceTokenizer::tokenize(const std::wstring& text) const {
-    std::vector<std::wstring> outputTokens;
-    for (auto& token : whitespaceTokenize(text)) {
-        if (token.size() > mMaxInputCharsPerWord) {
-            outputTokens.push_back(mUnkToken);
-        }
-        bool isBad = false;
-        size_t start = 0;
-        std::vector<std::wstring> subTokens;
-        while (start < token.size()) {
-            size_t end = token.size();
-            std::wstring curSubstr;
-            bool hasCurSubstr = false;
-            while (start < end) {
-                std::wstring substr = token.substr(start, end - start);
-                if (start > 0) substr = L"##" + substr;
-                if (mVocab->find(substr) != mVocab->end()) {
-                    curSubstr = substr;
-                    hasCurSubstr = true;
-                    break;
-                }
-                end--;
-            }
-            if (!hasCurSubstr) {
-                isBad = true;
-                break;
-            }
-            subTokens.push_back(curSubstr);
-            start = end;
-        }
-        if (isBad) outputTokens.push_back(mUnkToken);
-        else outputTokens.insert(outputTokens.end(), subTokens.begin(), subTokens.end());
-    }
-    return outputTokens;
-}
-
 
 class FullTokenizer {
 public:
@@ -236,24 +91,12 @@ private:
     WordpieceTokenizer mWordpieceTokenizer;
 };
 
-FullTokenizer::FullTokenizer(const std::string& vocabFile, bool doLowerCase) : 
-    mVocab(loadVocab(vocabFile)), 
-    mBasicTokenizer(BasicTokenizer(doLowerCase)),
-    mWordpieceTokenizer(WordpieceTokenizer(mVocab)) {
-    for (auto& v : *mVocab) mInvVocab[v.second] = v.first;
-}
-
-std::vector<std::wstring> FullTokenizer::tokenize(const std::string& text) const {
-    std::vector<std::wstring> splitTokens;
-    for (auto& token : mBasicTokenizer.tokenize(text))
-        for (auto& subToken : mWordpieceTokenizer.tokenize(token))  
-            splitTokens.push_back(subToken);
-    return splitTokens;
-}
-
 
 } // namespace ann
 } // namespace mlpack
+
+// Include implementation.
+#include "BERT_tokenizer_impl.hpp"
 
 
 #endif
